fix: Apply defaults to existing instances.
Previously if we saw a test instance with __typename already set,
we would stop and leave it as-is.

This was primarly to stop infinite recursion (i.e. author has a summary
that is required, which has an author that is required, which has a
summary...and repeat).

The __typename key existence was our signal that "this is an existing
factory instance, it doesn't need re-recreating".

However, this approach would break on factory calls that tried to stitch
together existing + slightly updated instances, b/c the existing
instance would bring in a __typename, so we would stop, but then the
"slightly updated" data might not pass through a factory, which would
mean it's __typename may not be set.

Which resulted in very opaque/old behavior from the Apollo Cache.

So, this fixes it, and instead we use an extra "all" set in our
existing cache param to see that "we've already touched this
instance".